Добавляет слайс user
